requirements.txt
# ─────────────────────────────────────────────
# BWA — Blog Writing Agent
# Backend Python Requirements
# ─────────────────────────────────────────────

# Web framework & server
fastapi==0.135.3
uvicorn==0.42.0
python-dotenv==1.2.1

# LangGraph / LangChain (agentic pipeline)
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-classic==1.0.1

# LLM providers
openai==2.16.0          # NVIDIA NIM compatible (used via base_url override)
google-genai==1.61.0    # Gemini image generation

# Web research
requests==2.32.5
aiohttp==3.13.3
httpx==0.28.1
httpx-sse==0.4.3
beautifulsoup4==4.14.3

# Data / validation
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
dataclasses-json==0.6.7
